backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.28.0
pycryptodome>=3.15.0
python-dotenv>=0.19.0
Pillow>=9.0.0
